Enforced semantic versioning from bash version checker.
As much as I looked in, it seems that there is no way to package together orb code and our python script, or for that matter any external code. If I'm right, the orb will only reference the external script relative to the config. However, this means that that external script required by orb would need to be added to every project that is importing the orb from the registry. Hence, I believe that we should stick to pure Bash script within the orb YAML file for now.